+++ Model/src/main/java/Monopoly/Player/Player.java
@@ -9,7 +9,7 @@ import java.util.Objects;
 import java.util.Random;
 
 public class Player extends Participant{
-    private int money;
+    private double money;
     private Pawn pawn;
     private Status isInJail;
     private int inJailTurn;
@@ -63,11 +63,11 @@ public class Player extends Participant{
         this.canExitJail = canExitJail;
     }
 
-    public int getMoney() {
+    public double getMoney() {
         return money;
     }
 
-    public void setMoney(int money) {
+    public void setMoney(double money) {
         this.money = money;
     }
